deconvolution layer init
