+++ src/plugins/intel_cpu/src/nodes/deconv.h
@@ -95,7 +95,7 @@ private:
     bool useACL = false;
     DeconvAttrs deconvAttrs;
 
-    Shape inShape;
+    Shape inShape, outShape;
 
     AttrPtr pAttr;
